WIP: Tag renaming based on --grammar-out
Currently does it by hooking into printSet function of GrammarWriter, so
the grammar is already loaded into vislcg3, processed and then rewritten
to file in changed form.

We want to allow changes from single tags to new sets, like

  (N)		TO	(np) OR (n)
  (Prop)	TO	(np)

or

  (n)		TO	(N) - (Prop)
  (np)		TO	(N Prop)

So for the first example, we could turn

  LIST foo = N adv;
  LIST bar = (N Prop) adv;
  LIST fie = (N Prop Sg Gen) adv;

into

  SET CG_RELABEL_1 = (np) OR (n) ;
  SET CG_RELABEL_2 = (np) ;
  SET foo = CG_RELABEL_1 OR (adv) ;
  SET bar = CG_RELABEL_1 + CG_RELABEL_2 OR (adv) ;
  SET fie = CG_RELABEL_1 + CG_RELABEL_2 + (Sg Gen) OR (adv) ;

by first outputting all the TO-parts as sets with generated names, then
as we output the tags of each LIST, we turn the LIST into a SET so that
we can use the + operator.

Caveats: Unified $$sets actually have to be "plain" list's, so anything
that's used in unification would have to make do with simple and
possibly wrong tag changes, e.g.

 LIST POS = N V ;

to

 LIST POS = np n vblex ;

TODO:
- Simple tag renaming on unified $$sets
- Should reading some input format into two tables (unifiable LIST vs
  SET version), currently just hardcoded test
- Should be a separate command, not happen during any --grammar-out ...
- COPY rules currently don't seem to work on the --grammar-out files
  even in the svn version
- SUBSTITUTE rules don't seem to work if the match is changed from a
  LIST to a SET